Remove LinkRequest struct
Moved the majority of its logic over to ResolvedLink
